cilium: add benchmark test class with base netperf test
Add a new class of tests for running benchmarks on the datapath. This
allowsjus to repeat performance test easily.

At first some basic netperf tests are created but these can be easily
extended. The output is reported in a perf.log file in the normal
test_results directory.

What about kubernetes-netperftest.sh? The kubernetes script is great
but gives us little control over how the netperf tests are executed
and what configurations are used from Cilium side. By writing our
own (and it doesn't take that much work) we get native ginkgo support
and full control which we can leverage for further testing.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>